+++ mathematica/SumsOfPowersViaBackwardFiniteDifferencesAndNewtonFormula.m
@@ -7,6 +7,9 @@ BackwardDifference::usage=""
 MultifoldSumOfPowersRecurrence::usage=""
 OrdinarySumsOfPowersViaBackwardDifferences::usage=""
 ValidateOrdinarySumsOfPowersViaBackwardDifferences::usage=""
+EulerianNumber::usage=""
+BackwardDifferencesInEulerianNumbers::usage=""
+ValidateBackwardDifferencesInEulerianNumbers::usage=""
 
 (*END: Definitions *)
 (* =========================================================================DOCS END=================================================================== *)
@@ -26,6 +29,11 @@ MultifoldSumOfPowersRecurrence[r_, n_, m_]:= n^m /; r==0;
 MultifoldSumOfPowersRecurrence[r_, n_, m_]:= Sum[MultifoldSumOfPowersRecurrence[r-1, k, m], {k, 1, n}] /; r>0;
 OrdinarySumsOfPowersViaBackwardDifferences[n_, m_, t_] := Sum[BackwardDifference[t, m, j] * ((-1)^j * Binomial[t, j+1] + Binomial[j-t+n, j+1]), {j, 0, m}];
 ValidateOrdinarySumsOfPowersViaBackwardDifferences[max_]:=Table[MultifoldSumOfPowersRecurrence[1, n, m] - OrdinarySumsOfPowersViaBackwardDifferences[n, m, t], {n, 0, max}, {m, 0, max}, {t, 0, max}]//Flatten
+EulerianNumber[0, 0] = 1;
+EulerianNumber[n_, k_] /; k < 0 || k >= n := 0;
+EulerianNumber[n_, k_] := EulerianNumber[n, k] = (k + 1) EulerianNumber[n - 1, k] + (n - k) EulerianNumber[n - 1, k - 1];
+BackwardDifferencesInEulerianNumbers[t_, m_, j_] := Sum[EulerianNumber[m, k] * Binomial[t+k-j, m-j], {k, 0, m}];
+ValidateBackwardDifferencesInEulerianNumbers[max_] := Table[BackwardDifference[t, m, j]-BackwardDifferencesInEulerianNumbers[t, m, j], {m, 0, max}, {j, 0, m}, {t, 0, max}] //Flatten
 (*END: Definitions *)
 End[ ]
 EndPackage[ ]