+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Mathematica programs for ordinary sums of powers
 - Mathematica documentation
 - Update Mathematica references
+- Formula for backward differences in Eulerian numbers (Worpitzky identity)
+- Mathematica for backward differences in Eulerian numbers (Worpitzky identity)
 
 ## [1.0.1] - 01-Jan-2026